@@ -98,17 +98,12 @@ where
             feature
                 .get_names()
                 .iter()
-                .map(|name| format!("bins_window{:.1}_offset{:.1}_{}", window, offset, name)),
+                .map(|name| format!("bins_window{window:.1}_offset{offset:.1}_{name}")),
         );
         self.properties
             .descriptions
             .extend(feature.get_descriptions().iter().map(|desc| {
-                format!(
-                    "{desc} for binned time-series with window {window} and offset {offset}",
-                    desc = desc,
-                    window = window,
-                    offset = offset,
-                )
+                format!("{desc} for binned time-series with window {window} and offset {offset}",)
             }));
         self.feature_extractor.add_feature(feature);
         self
